chore: mv ticket context and types out of HomePage.tsx
